Restructure
The order of steps changed. Now, the process is sequential: data generation, Bayes factor computation, and evaluation of power criterion are carried out for H1; after this, the sample size increases or the algorithm proceeds with H2.
Creation of two functions for sample size determination. One is the final binary search to stop the search of sample size, and another function for updating the sample size.