Include distinction of combined PMP for each hypothesis
There was no distinction between the different combined PMP when each hypothesis is true. Therefore, previous and current eta were zero all the time. In addition, it was unclear what must be returned as result.
Changed the name of the objects with combined PMPs  and added additional elements in the list that is returned.